Fix unicode issues and partially synced photo error
Unicode issue: my filesystem is UTF-8, so os.listdir returns unicode, but as
far as I see, titles got from online is UTF-8 - this must be converted to
unicode.

partially synced photo error: at filling rename_to_from_map, synced, but
online not existing photos are added to rename_to_from_map, which bails out
in the next if. Although the elif branch seems to cope with this case, too.